# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/users/me": {
      "patch": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/me/data": {
      "get": {
        "operationId": "UsersController_getMyData",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user personal data",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/me/delete-request": {
      "post": {
        "operationId": "UsersController_requestDeletion",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request account deletion",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/me/tos-agreement-latest": {
      "get": {
        "operationId": "UsersController_getMyLatestTosAgreement",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user latest TOS agreement",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/{id}/tos-agreements": {
      "get": {
        "operationId": "UsersController_getUserTosAgreements",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a user TOS agreement audit trail (admin only)",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get user by ID (public: id, name, createdAt)",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lendly API",
    "description": "Lendly backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          }
        }
      },
...
    }
  }
}